website/requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
psycopg2
PyJWT
python-multipart
